requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
databases
pydantic
python-dotenv
alembic
psycopg2